refactor: move AccessControl to own class
