+++ tsconfig.json
@@ -14,5 +14,11 @@
   },
   "include": [
     "src/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules",
+    "**/*.spec.ts",
+    "**/*.integ.ts",
+    "**/__fixtures__/*.*"
   ]
 }
